Refactor(api) : 임계치 관련 DTO 오타 및 타입 수정

+++ src/main/java/com/pooli/line/domain/dto/request/UpdateIndividualThresholdReqDto.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class UpdateIndividualThresholdReqDto {
 
     @Schema(description = "회선 ID", example = "1")
-    private Integer lineId;
+    private Long lineId;
 
     @NotNull @NotEmpty
     @Schema(description = "개인 데이터 임계치(MB)", example = "3000")
